+++ internal/service/federatedsettingsorgconfig/resource_federated_settings_connected_org_test.go
@@ -1,4 +1,4 @@
-package mongodbatlas_test
+package federatedsettingsorgconfig_test
 
 import (
 	"context"
