+++ internal/service/federatedsettingsorgrolemapping/data_source_federated_settings_org_role_mapping_test.go
@@ -1,4 +1,4 @@
-package mongodbatlas_test
+package federatedsettingsorgrolemapping_test
 
 import (
 	"fmt"
