+++ internal/service/federatedsettingsorgrolemapping/data_source_federated_settings_org_role_mappings.go
@@ -1,4 +1,4 @@
-package mongodbatlas
+package federatedsettingsorgrolemapping
 
 import (
 	"context"
@@ -12,7 +12,7 @@ import (
 	matlas "go.mongodb.org/atlas/mongodbatlas"
 )
 
-func DataSourceFederatedSettingsOrganizationRoleMappings() *schema.Resource {
+func PluralDataSource() *schema.Resource {
 	return &schema.Resource{
 		ReadContext: dataSourceMongoDBAtlasFederatedSettingsOrganizationRoleMappingsRead,
 		Schema: map[string]*schema.Schema{
